Use longs for entry counts everywhere

+++ store/src/main/scala/ionic/store/series/SeriesWriter.scala
@@ -23,7 +23,7 @@ object SeriesWriter {
       close()
   }
 
-  def writeMeta(dest: Directory, written: Int = 0, transferredFrom: String = "") {
+  def writeMeta(dest: Directory, written: Long = 0, transferredFrom: String = "") {
     val metaOut = dest.open("meta.avsc").overwrite()
     val encoder = EncoderFactory.get().jsonEncoder(SeriesMetadata.SCHEMA$, metaOut)
     val metaWriter: DatumWriter[SeriesMetadata] = new SpecificDatumWriter(SeriesMetadata.SCHEMA$)
